+++ datapipe_app/api_v1alpha2.py
@@ -249,7 +249,7 @@ def make_app(
 
     @app.get("/graph", response_model=models.GraphResponse)
     def get_graph() -> models.GraphResponse:
-        def table_response(table_name):
+        def table_response(table_name) -> models.TableResponse:
             tbl = catalog.get_datatable(ds, table_name)
 
             return models.TableResponse(
@@ -259,8 +259,8 @@ def make_app(
                 store_class=tbl.table_store.__class__.__name__,
             )
 
-        def pipeline_step_response(step: ComputeStep):
-            inputs = [i.name for i in step.input_dts]
+        def pipeline_step_response(step: ComputeStep) -> models.PipelineStepResponse:
+            inputs = [i.dt.name for i in step.input_dts]
             outputs = [i.name for i in step.output_dts]
 
             if isinstance(step, BaseBatchTransformStep):